desktop-app/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# AI Providers
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
httpx==0.26.0  # For Grok/xAI

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.3
python-multipart==0.0.9
pybreaker==1.2.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cors==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0  # For testing

# Development
black==24.2.0
ruff==0.2.1
mypy==1.8.0
bandit[toml]==1.7.7